+++ R/datasets.R
@@ -1,0 +1,7 @@
+#' Color palettes for the gsed data
+#'
+#' A list containing six color palettes for the gsed data.
+#'
+#' @docType data
+#' @format A named \code{list} with vectors.
+"gsed_palettes"